+++ src/GafferImage/OpenColorIOAlgo.cpp
@@ -60,14 +60,22 @@ const string g_sceneLinearString( OCIO_NAMESPACE::ROLE_SCENE_LINEAR );
 IECorePreview::LRUCache<std::string, OCIO_NAMESPACE::ConstConfigRcPtr> g_configCache(
 	[] ( const std::string &fileName, size_t &cost, const IECore::Canceller *canceller ) {
 		cost = 1;
+		OCIO_NAMESPACE::ConstConfigRcPtr config;
 		if( fileName.empty() )
 		{
-			return OCIO_NAMESPACE::Config::CreateFromEnv();
+			config = OCIO_NAMESPACE::Config::CreateFromEnv();
 		}
 		else
 		{
-			return OCIO_NAMESPACE::Config::CreateFromFile( fileName.c_str() );
+			config = OCIO_NAMESPACE::Config::CreateFromFile( fileName.c_str() );
 		}
+		// Various config queries such as `getDefaultDisplay()` are not
+		// threadsafe, because they do an unguarded lazy-initialisation of a
+		// data structure within the config. Force that initialisation now
+		// while it's protected by our LRUCache mutex, freeing OpenColorIOTransform
+		// nodes from worrying about it later.
+		config->getDefaultDisplay();
+		return config;
 	},
 	1000
 );